+++ web/src/contexts/message/index.jsx
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+
+/**
+ * messages, dispatch
+ */
+export const MessageContext = createContext({
+    messages: [],
+    dispatch: () => { },
+});
